Move useProfileTab out of pages/ to helpers/hooks
